chore: move mute types to notify package
Move `Muter`, `MuteFunc`, `Mutes` and `TimeMuter` to notify package.
